refactoring | #41 | create FileManager component
- Replaced logic to work for file system in own component
