+++ packages/cli-ui/src/components/FileManager/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FileManager'
