+++ packages/cli-ui/src/components/index.ts
@@ -10,3 +10,4 @@ export { default as Toolbar } from './Toolbar'
 export { default as KillPort } from './KillPort'
 export { default as CheckBoxTheme } from './CheckBoxTheme'
 export { default as Empty } from './Empty'
+export { default as FileManager } from './FileManager'
